update chapter 04

+++ include/helperUtils.h
@@ -21,7 +21,7 @@ inline double randomDouble(double min, double max) {
     return min + (max - min) * randomDouble();
 }
 
-inline double randomInt(int min, int max) {
+inline int randomInt(int min, int max) {
     // Returns a random integer in [min,max].
     return static_cast<int>(randomDouble(min, max + 1.0));
 }
